+++ usr/src/cmd/cmd-inet/usr.bin/pppd/utils.c
@@ -30,6 +30,9 @@
  * IMPLIED WARRANTIES, INCLUDING, WITHOUT LIMITATION, THE IMPLIED
  * WARRANTIES OF MERCHANTIBILITY AND FITNESS FOR A PARTICULAR PURPOSE.
  */
+/*
+ * Copyright 2017 Hayashi Naoyuki
+ */
 
 #ifdef __linux__
 #define	_GNU_SOURCE
@@ -509,7 +512,7 @@ log_packet(p, len, prefix, level)
 {
     (void) strlcpy(line, prefix, sizeof(line));
     linep = line + strlen(line);
-    format_packet(p, len, pr_log, (void *)level);
+    format_packet(p, len, pr_log, (void *)(intptr_t)level);
     if (linep != line)
 	syslog(level, "%s", line);
 }
@@ -578,7 +581,7 @@ pr_log __V((void *arg, const char *fmt, ...))
     va_end(pvar);
 
     if (linep + n + 1 > line + sizeof(line)) {
-	syslog((int)arg, "%s", line);
+	syslog((intptr_t)arg, "%s", line);
 	linep = line;
     }
     (void) strlcpy(linep, buf, line + sizeof(line) - linep);
